Refactored Types Folder
