+++ types/common.ts
@@ -1,0 +1,22 @@
+// Common Types
+export interface SelectOption {
+  label: string;
+  value: string;
+}
+
+export interface TableColumn<T = any> {
+  key: keyof T;
+  label: string;
+  sortable?: boolean;
+  render?: (value: any, row: T) => React.ReactNode;
+}
+
+export interface FormField {
+  name: string;
+  label: string;
+  type: 'text' | 'email' | 'password' | 'select' | 'textarea' | 'file';
+  required?: boolean;
+  placeholder?: string;
+  options?: SelectOption[];
+  validation?: any;
+}
